Remove Event class from REST object hierarchy
